+++ 01.javascript/todoapp-router/assets/js/pages/list/TodoList.js
@@ -3,6 +3,7 @@ import Header from "../../layout/Header.js";
 import Footer from "../../layout/Footer.js";
 import { linkTo } from "../../Router.js";
 import Button from "../../layout/Button.js";
+import TodoListItem from "./TodoListItem.js";
 
 const TodoList = async function () {
   const page = document.createElement("div");
@@ -56,57 +57,12 @@ const TodoList = async function () {
     const ul = document.createElement("ul");
     ul.setAttribute("class", "todolist");
     // ul.setAttribute = ("data-deadline", `${item.createdAt}`);
-
-    const checkboxes = [];
+    const checkboxList = [];
     response.data?.items.forEach((item) => {
       console.log(item);
       /* todoItem 초기렌더링 */
-      const li = document.createElement("li");
-      li.setAttribute("class", "todolist__item");
 
-      // TODO: 클릭하면 important 속성 변경
-      /* todoItem 중요버튼 */
-      const importantButton = document.createElement("button");
-      importantButton.setAttribute("class", `todolist__item--important-button ${!item.important ? null : "fill"}`);
-      importantButton.addEventListener("click", () => {
-        importantButton.classList.toggle("fill");
-      });
-
-      // TODO: 클릭하면 삭제
-      /* todoItem 삭제 버튼 */
-      const deleteButton = document.createElement("a");
-      deleteButton.setAttribute("class", "todolist__item--delete-button");
-      deleteButton.setAttribute("href", `?_id=${item._id}`);
-
-      /* todoItem 상세보기 모달링크 */
-      const todoInfoLink = document.createElement("a");
-      todoInfoLink.setAttribute("href", `info?_id=${item._id}`);
-      todoInfoLink.setAttribute("class", "todolist__item--todo-info-link");
-      todoInfoLink.setAttribute("for", `${item.title}`);
-      const title = document.createTextNode(item.title);
-
-      // TODO: 클릭하면 done 값 변경
-      /* todoItem 체크박스 */
-      const checkbox = document.createElement("input");
-      checkbox.setAttribute("type", "checkbox");
-      checkbox.setAttribute("class", "todolist__item--checkbox");
-      checkbox.setAttribute("id", `${item.title}`);
-      if (item.done) {
-        checkbox.setAttribute("checked", true);
-        /* item이 done일때 취소선 스타일링 */
-        todoInfoLink.style.textDecoration = "line-through";
-      }
-      checkboxes.push(checkbox);
-      todoInfoLink.addEventListener("click", function (event) {
-        event.preventDefault();
-        linkTo(todoInfoLink.getAttribute("href"));
-      });
-
-      todoInfoLink.appendChild(title);
-      li.appendChild(importantButton);
-      li.appendChild(checkbox);
-      li.appendChild(todoInfoLink);
-      li.appendChild(deleteButton);
+      const li = TodoListItem(item, checkboxList);
       if (item.important) {
         importantList.appendChild(li);
       } else {
@@ -120,11 +76,11 @@ const TodoList = async function () {
     contents.appendChild(importantList);
     contents.appendChild(ul);
 
-    /* 전체완료 버튼 토글링 */
+    /* 전체완료 체크박스 토글링 */
     let toggleCompletAll = Number(completedAll.dataset.done);
     completedAll.addEventListener("click", () => {
       toggleCompletAll = !toggleCompletAll;
-      checkboxes.forEach((checkbox) => (checkbox.checked = toggleCompletAll));
+      checkboxList.forEach((checkbox) => (checkbox.checked = toggleCompletAll));
       // checkboxes
     });
